+++ src/facades.php
@@ -81,6 +81,8 @@ const isnull = '\Siler\Functional\isnull';
 const concat = '\Siler\Functional\concat';
 const lazy = '\Siler\Functional\lazy';
 const call = '\Siler\Functional\call';
+const map = '\Siler\Functional\map';
+const lmap = '\Siler\Functional\lmap';
 
 
 namespace Siler\Functional\Monad;
@@ -243,6 +245,7 @@ const bearer = '\Siler\Swoole\bearer';
 const http_server_port = '\Siler\Swoole\http_server_port';
 const graphql_handler = '\Siler\Swoole\graphql_handler';
 const redirect = '\Siler\Swoole\redirect';
+const middleware = '\Siler\Swoole\middleware';
 
 
 namespace Siler\Twig;